fix: create default wallet for auto created `super_user`
